Factorize code between autocontainer and hashedcontainer
